move data to package
