+++ training_templates/data_utils.py
@@ -1,4 +1,14 @@
+import os
+
+import pandas as pd
 from pyspark.sql import SparkSession
+from pyspark.sql.types import (
+    FloatType,
+    IntegerType,
+    StringType,
+    StructField,
+    StructType,
+)
 
 
 def train_test_split(
@@ -56,3 +66,61 @@ def train_test_split(
     )
 
     return (train_output_table_name, test_output_table_name)
+
+
+def sample_pandas_dataframe():
+    dtype = {
+        "PassengerId": "object",
+        "Fare": "float64",
+        "Embarked": "object",
+        "Pclass": "object",
+        "Parch": "object",
+        "TicketChars": "object",
+        "CabinChar": "object",
+        "CabinMulti": "object",
+        "FareRounded": "float64",
+        "Name": "object",
+        "Sex": "object",
+        "Age": "float64",
+        "SibSp": "object",
+        "NamePrefix": "object",
+        "NameSecondary": "object",
+        "NameMultiple": "object",
+        "Survived": "int32",
+    }
+
+    current_dir = os.path.dirname(__file__)
+    data_path = f"{current_dir}/data/sample_features.csv"
+
+    sample_dataframe = pd.read_csv(data_path, header=0, dtype=dtype)
+    return sample_dataframe
+
+
+def sample_spark_dataframe():
+    spark = SparkSession.builder.getOrCreate()
+    sample_pandas_df = sample_pandas_dataframe()
+
+    spark_schema = StructType(
+        [
+            StructField("PassengerId", StringType(), True),
+            StructField("Fare", FloatType(), True),
+            StructField("Embarked", StringType(), True),
+            StructField("Pclass", StringType(), True),
+            StructField("Parch", StringType(), True),
+            StructField("TicketChars", StringType(), True),
+            StructField("CabinChar", StringType(), True),
+            StructField("CabinMulti", StringType(), True),
+            StructField("FareRounded", FloatType(), True),
+            StructField("Name", StringType(), True),
+            StructField("Sex", StringType(), True),
+            StructField("Age", FloatType(), True),
+            StructField("SibSp", StringType(), True),
+            StructField("NamePrefix", StringType(), True),
+            StructField("NameSecondary", StringType(), True),
+            StructField("NameMultiple", StringType(), True),
+            StructField("Survived", IntegerType(), True),
+        ]
+    )
+
+    sample_dataframe = spark.createDataFrame(sample_pandas_df, schema=spark_schema)
+    return sample_dataframe
